Clean code.
