+++ java/hive/src/main/java/io/v6d/hive/ql/io/VineyardInputFormat.java
@@ -14,295 +14,78 @@
  */
 package io.v6d.hive.ql.io;
 
-import io.v6d.core.common.util.ObjectID;
-import org.apache.arrow.memory.BufferAllocator;
-import io.v6d.core.common.util.VineyardException;
-import org.apache.arrow.memory.RootAllocator;
 import io.v6d.core.client.IPCClient;
 import io.v6d.core.client.ds.ObjectFactory;
-import io.v6d.core.client.ds.ObjectMeta;
-import io.v6d.modules.basic.arrow.TableBuilder;
-import io.v6d.modules.basic.arrow.SchemaBuilder;
+import io.v6d.core.common.util.ObjectID;
 import io.v6d.modules.basic.arrow.Arrow;
-import io.v6d.modules.basic.arrow.RecordBatchBuilder;
 import io.v6d.modules.basic.arrow.Table;
 
-import org.apache.hadoop.hive.llap.FieldDesc;
-import org.apache.hadoop.hive.llap.Row;
-import org.apache.hadoop.hive.llap.LlapArrowRowRecordReader;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+import org.apache.hadoop.hive.ql.exec.Utilities;
+import org.apache.hadoop.hive.ql.exec.vector.*;
 import org.apache.hadoop.hive.ql.io.HiveInputFormat;
-import org.apache.hadoop.hive.ql.io.arrow.ArrowWrapperWritable;
-import org.apache.hadoop.hive.serde2.ColumnProjectionUtils;
-import org.apache.hadoop.hive.serde2.typeinfo.PrimitiveTypeInfo;
-import org.apache.hadoop.hive.serde2.typeinfo.TypeInfoFactory;
-import org.apache.hadoop.io.*;
-import org.apache.hadoop.mapred.*;
+import org.apache.hadoop.io.NullWritable;
+import org.apache.hadoop.mapred.FileInputFormat;
+import org.apache.hadoop.mapred.InputSplit;
+import org.apache.hadoop.mapred.JobConf;
+import org.apache.hadoop.mapred.RecordReader;
+import org.apache.hadoop.mapred.Reporter;
+
+import org.apache.arrow.vector.*;
+import org.apache.arrow.vector.VectorSchemaRoot;
+import org.apache.arrow.vector.types.pojo.*;
+import org.apache.hadoop.fs.Path;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.google.flatbuffers.IntVector;
-
-import java.io.IOException;
-
-import org.apache.arrow.vector.VectorSchemaRoot;
-import org.apache.arrow.vector.complex.NonNullableStructVector;
-import org.apache.arrow.vector.types.pojo.*;
-import org.apache.hadoop.fs.Path;
-import org.apache.arrow.vector.FieldVector;
-import java.util.ArrayList;
-import java.util.List;
-import java.io.DataOutput;
-import java.io.DataInput;
-
-import lombok.val;
-// how to connect vineyard
-
-// check phase:
-// 1. check if vineyard is running
-// 2. check if file exist in vineyard
-// 3. if exist, read file from vineyard
-// 4. if not exist, read file from hdfs and write to vineyard
-//
-// We do not split the file at present.
-
-class RowWritable implements Writable {
-    private Row row;
-    private int columnCount = 0;
-
-    public RowWritable(final Row row) {
-        this.row = row;
-        this.columnCount = row.getSchema().getColumns().size();
-    }
-
-    public Row getRow() {
-        return row;
-    }
-
-    public Object[] getValues() {
-        Object[] values = new Object[this.columnCount];
-        for (int i = 0; i < this.columnCount; i++) {
-            // FIXME(tao): can be avoid the transformation to `Writable`?
-            values[i] = makeWritable(this.row.getValue(i));
-        }
-        return values;
-    }
+public class VineyardInputFormat extends HiveInputFormat<NullWritable, VectorizedRowBatch> implements VectorizedInputFormatInterface {
 
     @Override
-    public void write(DataOutput out) throws IOException {
-    }
-
-    @Override
-    public void readFields(DataInput in) throws IOException {
-    }
-
-    private BooleanWritable makeWritable(boolean value) {
-        return new BooleanWritable(value);
-    }
-
-    private IntWritable makeWritable(int value) {
-        return new IntWritable(value);
-    }
-
-    private LongWritable makeWritable(long value) {
-        return new LongWritable(value);
-    }
-
-    private FloatWritable makeWritable(float value) {
-        return new FloatWritable(value);
-    }
-
-    private DoubleWritable makeWritable(double value) {
-        return new DoubleWritable(value);
-    }
-
-    private Text makeWritable(String value) {
-        return new Text(value);
-    }
-
-    private Object makeWritable(Object value) {
-        if (value == null) {
-            return null;
-        }
-        if (value instanceof Boolean) {
-            return makeWritable((boolean) value);
-        }
-        if (value instanceof Integer) {
-            return makeWritable((int) value);
-        }
-        if (value instanceof Long) {
-            return makeWritable((long) value);
-        }
-        if (value instanceof Float) {
-            return makeWritable((float) value);
-        }
-        if (value instanceof Double) {
-            return makeWritable((double) value);
-        }
-        if (value instanceof String) {
-            return makeWritable((String) value);
-        }
-        return value;
-    }
-}
-
-class VineyardArrowRowRecordReader implements RecordReader<NullWritable, RowWritable> {
-    private LlapArrowRowRecordReader reader;
-
-    public VineyardArrowRowRecordReader(JobConf job, org.apache.hadoop.hive.llap.Schema schema, RecordReader<NullWritable, ArrowWrapperWritable> reader) throws IOException {
-        this.reader = new LlapArrowRowRecordReader(job, schema, reader);
-    }
-
-    @Override
-    public boolean next(NullWritable key, RowWritable value) throws IOException {
-        if (this.reader.next(key, value.getRow())) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public NullWritable createKey() {
-        return NullWritable.get();
-    }
-
-    @Override
-    public RowWritable createValue() {
-        return new RowWritable(new Row(this.reader.getSchema()));
-    }
-
-    @Override
-    public long getPos() throws IOException {
-        return 0;
-    }
-
-    @Override
-    public void close() throws IOException {
-        this.reader.close();
-    }
-
-    @Override
-    public float getProgress() throws IOException {
-        return 0;
-    }
-}
-
-public class VineyardInputFormat extends HiveInputFormat<NullWritable, RowWritable> {
-
-    @Override
-    public RecordReader<NullWritable, RowWritable>
+    public RecordReader<NullWritable, VectorizedRowBatch>
     getRecordReader(InputSplit genericSplit, JobConf job, Reporter reporter)
-        throws IOException {
+            throws IOException {
         reporter.setStatus(genericSplit.toString());
-        System.out.printf("--------+creating vineyard record reader\n");
-        System.out.println("split class:" + genericSplit.getClass().getName());
-        return new VineyardArrowRowRecordReader(job, this.makeDummySchema(), new VineyardRecordReader(job, (VineyardSplit) genericSplit));
+        return new VineyardBatchRecordReader(job, (VineyardSplit) genericSplit);
+    }
+
+    @Override
+    public VectorizedSupport.Support[] getSupportedFeatures() {
+        return new VectorizedSupport.Support[] {VectorizedSupport.Support.DECIMAL_64};
     }
 
     @Override
     public InputSplit[] getSplits(JobConf job, int numSplits) throws IOException {
-        System.out.println("--------+creating vineyard input split. Num:" + numSplits);
-        val columnIds = ColumnProjectionUtils.getReadColumnIDs(job);
-        System.out.printf("columnIds: %s\n", columnIds);
-        val columnNames = ColumnProjectionUtils.getReadColumnNames(job);
-        System.out.printf("columnNames: %s\n", columnNames);
         List<InputSplit> splits = new ArrayList<InputSplit>();
         Path path = FileInputFormat.getInputPaths(job)[0];
-        System.out.println("path:" + path);
         // fill splits
         VineyardSplit vineyardSplit = new VineyardSplit(path, 0, 0, job);
         splits.add(vineyardSplit);
         return splits.toArray(new VineyardSplit[splits.size()]);
     }
-
-    private org.apache.hadoop.hive.llap.Schema makeDummySchema() {
-        // Field field1 = new Field("field_1", FieldType.nullable(new ArrowType.Int(32, true)), null);//Arrow.makeField("field_1", Arrow.FieldType.Int);
-        // Field field2 = new Field("field_2", FieldType.nullable(new ArrowType.Int(32, true)), null);//Arrow.makeField("field_2", Arrow.FieldType.Int);
-        // List<Field> fields = new ArrayList<Field>();
-        // fields.add(field1);
-        // fields.add(field2);
-        // return new Schema(fields);
-        FieldDesc field1 = new FieldDesc("field_1", TypeInfoFactory.getPrimitiveTypeInfo("int"));
-        FieldDesc field2 = new FieldDesc("field_2", TypeInfoFactory.getPrimitiveTypeInfo("int"));
-        List<FieldDesc> fields = new ArrayList<FieldDesc>();
-        fields.add(field1);
-        fields.add(field2);
-        return new org.apache.hadoop.hive.llap.Schema(fields);
-    }
 }
 
-class VineyardSplit extends FileSplit {
-    String customPath;
-
-    protected VineyardSplit() {
-        super();
-    }
-
-    public VineyardSplit(Path file, long start, long length, JobConf conf) {
-        super(file, start, length, (String[]) null);
-    }
-
-    public VineyardSplit(Path file, long start, long length, String[] hosts) {
-        super(file, start, length, hosts);
-    }
-
-    @Override
-    public Path getPath() {
-        return super.getPath();
-    }
-
-    @Override
-    public long getLength() {
-        return 0;
-    }
-
-    @Override
-    public String[] getLocations() throws IOException {
-        return new String[0];
-    }
-
-    @Override
-    public void readFields(DataInput in) throws IOException {
-        super.readFields(in);
-    }
-
-    @Override
-    public void write(DataOutput out) throws IOException {
-        super.write(out);
-    }
-}
-
-class VineyardRecordReader implements RecordReader<NullWritable, ArrowWrapperWritable> {
-    private static Logger logger = LoggerFactory.getLogger(VineyardRecordReader.class);
+class VineyardBatchRecordReader implements RecordReader<NullWritable, VectorizedRowBatch> {
+    private static Logger logger = LoggerFactory.getLogger(VineyardBatchRecordReader.class);
 
     // vineyard field
-    private static IPCClient client;
+    private IPCClient client;
     private String tableName;
     private Boolean tableNameValid = false;
-    private VectorSchemaRoot vectorSchemaRoot = getSchemaRoot();
+    private VectorSchemaRoot vectorSchemaRoot;
+    private VectorizedRowBatchCtx ctx;
+    private Table table;
+    private int recordBatchIndex = 0;
+    private int recordBatchInnerIndex = 0;
 
-    // for test
-    private long tableID;
-    private TableBuilder tableBuilder;
-    private SchemaBuilder schemaBuilder;
-    private List<RecordBatchBuilder> recordBatchBuilders;
-    private static RecordBatchBuilder recordBatchBuilder;
-
-    VineyardRecordReader(JobConf job, VineyardSplit split) {
-        System.out.printf("--------+creating vineyard record reader\n");
-        // throw new RuntimeException("mapred record reader: unimplemented");
-        // reader = new LineRecordReader(job, split);
-        System.out.println("Path:" + split.getPath());
+    VineyardBatchRecordReader(JobConf job, VineyardSplit split) {
         String path = split.getPath().toString();
         int index = path.lastIndexOf("/");
         tableName = path.substring(index + 1);
         tableNameValid = true;
-        System.out.println("Table name:" + tableName);
-
-        val columnIds = ColumnProjectionUtils.getReadColumnIDs(job);
-        System.out.printf("columnIds: %s\n", columnIds);
-        val columnNames = ColumnProjectionUtils.getReadColumnNames(job);
-        System.out.printf("columnNames: %s\n", columnNames);
 
         // connect to vineyard
         if (client == null) {
@@ -322,18 +105,13 @@ class VineyardRecordReader implements RecordReader<NullWritable, ArrowWrapperWri
             System.out.println("Hello, vineyard!");
         }
 
-        try {
-            prepare();
-        } catch (Exception e) {
-            System.out.println(e.getMessage());
-        }
         Arrow.instantiate();
+        ctx = Utilities.getVectorizedRowBatchCtx(job);
     }
 
     @Override
     public void close() throws IOException {
-        System.out.printf("--------closing\n");
-        if(client.connected()) {
+        if(client != null && client.connected()) {
             client.disconnect();
             System.out.println("Bye, vineyard!");
         }
@@ -341,164 +119,138 @@ class VineyardRecordReader implements RecordReader<NullWritable, ArrowWrapperWri
 
     @Override
     public NullWritable createKey() {
-        System.out.printf("--------creating key\n");
         return null;
     }
 
     @Override
-    public ArrowWrapperWritable createValue() {
-        System.out.printf("++++++++creating value\n");
-        return new ArrowWrapperWritable(vectorSchemaRoot , Arrow.default_allocator, NonNullableStructVector.empty(tableName, Arrow.default_allocator));
+    public VectorizedRowBatch createValue() {
+        return ctx.createVectorizedRowBatch();
     }
 
     @Override
     public long getPos() throws IOException {
-        System.out.printf("+++++++get pos\n");
         return 0;
     }
 
     @Override
     public float getProgress() throws IOException {
-        System.out.printf("++++++++get progress\n");
         return 0;
     }
 
+    /**
+     * Copy data from the Arrow RecordBatch to the VectorizedRowBatch.
+     *
+     * @param recordBatch the Arrow RecordBatch to copy from
+     * @param batch the VectorizedRowBatch to copy to
+     */
+    private void arrowToVectorizedRowBatch(VectorizedRowBatch batch) {
+        // batch.numCols = recordBatch.getFieldVectors().size();
+        int remaining = vectorSchemaRoot.getRowCount() - recordBatchInnerIndex;
+
+        /*
+         * When we use SQL with condition to query data, the recordBatchSize may be large than VectorizedRowBatch.DEFAULT_SIZE, 
+         * which will cause ArrayIndexOutOfBoundsException. Because hive use a temporary array to store the data of
+         * VectorizedRowBatch when process condition query, and the size of the array is VectorizedRowBatch.DEFAULT_SIZE
+         * (VectorFilterOperator.java: 102).
+         * So we need to limit the recordBatchSize to VectorizedRowBatch.DEFAULT_SIZE if the recordBatchSize is larger than
+         * VectorizedRowBatch.DEFAULT_SIZE.
+         */
+        int recordBatchSize = remaining <= VectorizedRowBatch.DEFAULT_SIZE ? remaining : VectorizedRowBatch.DEFAULT_SIZE;
+        batch.size = recordBatchSize;
+
+        for (int i = 0; i < vectorSchemaRoot.getSchema().getFields().size(); i++) {
+            Field field = vectorSchemaRoot.getSchema().getFields().get(i);
+            if (field.getType().equals(Arrow.Type.Boolean)) {
+                LongColumnVector vector = new LongColumnVector(batch.size);
+                BitVector bitVector = (BitVector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.vector[k] = bitVector.get(k + recordBatchInnerIndex);
+                }
+                batch.cols[i] = vector;
+            } else if (field.getType().equals(Arrow.Type.Int)) {
+                LongColumnVector vector = new LongColumnVector(batch.size);
+                IntVector intVector = (IntVector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.vector[k] = intVector.get(k + recordBatchInnerIndex);
+                }
+                batch.cols[i] = vector;
+            } else if (field.getType().equals(Arrow.Type.Int64)) {
+                LongColumnVector vector = new LongColumnVector(batch.size);
+                BigIntVector bigIntVector = (BigIntVector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.vector[k] = bigIntVector.get(k + recordBatchInnerIndex);
+                }
+                batch.cols[i] = vector;
+            } else if (field.getType().equals(Arrow.Type.Float)) {
+                DoubleColumnVector vector = new DoubleColumnVector(batch.size);
+                Float4Vector float4Vector = (Float4Vector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.vector[k] = float4Vector.get(k + recordBatchInnerIndex);
+                }
+                batch.cols[i] = vector;
+            } else if (field.getType().equals(Arrow.Type.Double)) {
+                DoubleColumnVector vector = new DoubleColumnVector(batch.size);
+                Float8Vector float8Vector = (Float8Vector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.vector[k] = float8Vector.get(k + recordBatchInnerIndex);
+                }
+                batch.cols[i] = vector;
+            } else if (field.getType().equals(Arrow.Type.LargeVarChar)) {
+                BytesColumnVector vector = new BytesColumnVector(batch.size);
+                LargeVarCharVector largeVarCharVector = (LargeVarCharVector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.isNull[k] = false;
+                    vector.setRef(k, largeVarCharVector.get(k + recordBatchInnerIndex), 0, largeVarCharVector.get(k + recordBatchInnerIndex).length);
+                }
+                batch.cols[i] = vector;
+            } else if (field.getType().equals(Arrow.Type.VarChar)) {
+                BytesColumnVector vector = new BytesColumnVector(batch.size);
+                vector.init();
+                VarCharVector varCharVector = (VarCharVector) vectorSchemaRoot.getFieldVectors().get(i);
+                for (int k = 0; k < batch.size; k++) {
+                    vector.isNull[k] = false;
+                    vector.setRef(k, varCharVector.get(k + recordBatchInnerIndex), 0, varCharVector.get(k + recordBatchInnerIndex).length);
+                }
+                batch.cols[i] = vector;
+            } else {
+                System.out.println("array builder for type " + field.getType() + " is not supported");
+                throw new UnsupportedOperationException("array builder for type " + field.getType() + " is not supported");
+            }
+        }
+        recordBatchInnerIndex += recordBatchSize;
+        if (recordBatchInnerIndex >= vectorSchemaRoot.getRowCount()) {
+            recordBatchInnerIndex = 0;
+            recordBatchIndex++;
+            vectorSchemaRoot = null;
+        }
+    }
+
     @Override
-    public boolean next(NullWritable key, ArrowWrapperWritable value) throws IOException {
-        System.out.printf("+++++++++next\n");
+    public boolean next(NullWritable key, VectorizedRowBatch value) throws IOException {
         if (tableNameValid) {
-            // read(tableName);
-            // if exist, create vectorSchema root set data and return true;
-            // return false;
-            vectorSchemaRoot = getSchemaRoot();
-            if (vectorSchemaRoot == null) {
-                System.out.println(tableName + " not exist!");
+            if (table == null) {
+                try {
+                    ObjectID objectID = client.getName(tableName, false);
+                    table = (Table) ObjectFactory.getFactory().resolve(client.getMetaData(objectID));
+                } catch (Exception e) {
+                    System.out.println("Get objectID failed.");
+                    return false;
+                }
+            }
+            if (recordBatchIndex >= table.getBatches().size()) {
                 return false;
             }
-            System.out.println("Get schema root succeed!");
-            value.setVectorSchemaRoot(vectorSchemaRoot);
-            NonNullableStructVector vector = value.getRootVector();
-            vector.initializeChildrenFromFields(vectorSchemaRoot.getSchema().getFields());
-            vector.setValueCount(vectorSchemaRoot.getRowCount());
-            List<FieldVector> fieldVectors = vector.getChildrenFromFields();
-            System.out.println("fieldVectors size: " + fieldVectors.size());
-            for (int i = 0; i < fieldVectors.size(); i++) {
-                System.out.println("fieldVectors[" + i + "] size: " + fieldVectors.get(i).getValueCount());
-                for (int j = 0; j < fieldVectors.get(i).getValueCount(); j++) {
-                    ((org.apache.arrow.vector.IntVector)(fieldVectors.get(i))).set(j, (int)(vectorSchemaRoot.getFieldVectors().get(i).getObject(j)));
-                    System.out.println(fieldVectors.get(i).getObject(j));
+            if (vectorSchemaRoot == null) {
+                vectorSchemaRoot = table.getArrowBatch(recordBatchIndex);
+            }
+            if (vectorSchemaRoot == null) {
+                System.out.println("Get vector schema root failed.");
+                return false;
+            }
 
-                }
-                System.out.println("null count:" + ((org.apache.arrow.vector.IntVector)fieldVectors.get(i)).getNullCount());
-            }
-            System.out.println("========================");
-            if (value.getVectorSchemaRoot() != null) {
-                System.out.println("Set vectorSchemaRoot succeed!");
-            }
-            tableNameValid = false;
+            this.arrowToVectorizedRowBatch(value);
             return true;
         }
         return false;
     }
-
-    private VectorSchemaRoot getSchemaRoot() {
-        // System.out.println("Get table from vineyard.");
-        // if (tableNameValid == false) {
-        //     return null;
-        // }
-
-        // try {
-        //     ObjectID objectID = client.getName(tableName);
-        //     System.out.println("ObjectID: " + objectID.value());
-        //     ObjectMeta meta = client.getMetaData(objectID);
-        //     if (meta == null) {
-        //         System.out.println("Get meta failed.");
-        //     } else {
-        //         System.out.println("Get meta succeed. Meta id:" + meta.getId().value());
-        //     }
-        //     Table table = (Table) ObjectFactory.getFactory().resolve(meta);
-        //     if (table == null) {
-        //         System.out.println("Create table from factory failed!");
-        //     } else {
-        //         System.out.println("Create table from factory succeed!");
-        //     }
-        //     // TBD: support more than one batch
-        //     tableNameValid = false;
-        //     return table.getArrowBatch(0);
-        // } catch (Exception e) {
-        //     System.out.println("Get table failed!");
-        //     System.out.println(e.getMessage());
-        // }
-        BufferAllocator allocator = new RootAllocator();
-        Field field1 = new Field("field_1", FieldType.nullable(new ArrowType.Int(32, true)), null);//Arrow.makeField("field_1", Arrow.FieldType.Int);
-        Field field2 = new Field("field_2", FieldType.nullable(new ArrowType.Int(32, true)), null);//Arrow.makeField("field_2", Arrow.FieldType.Int);
-        List<Field> fields = new ArrayList<Field>();
-        fields.add(field1);
-        fields.add(field2);
-        // org.apache.arrow.vector.IntVector vector1 = new org.apache.arrow.vector.IntVector(field1, Arrow.default_allocator);
-        // org.apache.arrow.vector.IntVector vector2 = new org.apache.arrow.vector.IntVector(field2, Arrow.default_allocator);
-        // vector1.allocateNew(5);
-        // vector2.allocateNew(5);
-        // for (int i = 0; i < 5; i++) {
-        //     vector1.set(i, i);
-        //     vector2.set(i, i + 1);
-        // }
-        // List<FieldVector> fieldVectors = new ArrayList<FieldVector>();
-        // fieldVectors.add(vector1);
-        // fieldVectors.add(vector2);
-
-        // VectorSchemaRoot vectorSchemaRoot = new VectorSchemaRoot(fieldVectors);
-        Schema schema = new Schema(fields);
-        VectorSchemaRoot vectorSchemaRoot = VectorSchemaRoot.create(schema, allocator);
-        vectorSchemaRoot.setRowCount(5);
-        org.apache.arrow.vector.IntVector v1 = (org.apache.arrow.vector.IntVector) vectorSchemaRoot.getVector("field_1");
-        org.apache.arrow.vector.IntVector v2 = (org.apache.arrow.vector.IntVector) vectorSchemaRoot.getVector("field_2");
-        v1.allocateNew(5);
-        v1.setValueCount(5);
-        v2.allocateNew(5);
-        v2.setValueCount(5);
-        for (int i = 0; i < 5; i++) {
-            v1.set(i, i);
-            v2.set(i, i + 1);
-        }
-        // tableNameValid = false;
-        return vectorSchemaRoot;
-    }
-
-    private void prepare() throws VineyardException{
-        System.out.println("Prepare data");
-        SchemaBuilder schemaBuilder = new SchemaBuilder();
-        recordBatchBuilders = new ArrayList<RecordBatchBuilder>();
-        Field field1 = Arrow.makeField("field_1", Arrow.FieldType.Int);
-        Field field2 = Arrow.makeField("field_2", Arrow.FieldType.Int);
-        schemaBuilder.addField(field1);
-        schemaBuilder.addField(field2);
-        recordBatchBuilder = new RecordBatchBuilder(client, 5);
-        recordBatchBuilder.addField(field1);
-        recordBatchBuilder.addField(field2);
-        recordBatchBuilder.finishSchema(client);
-
-        val column1 = recordBatchBuilder.getColumnBuilder(0);
-        column1.setInt(0, 1);
-        column1.setInt(1, 2);
-        column1.setInt(2, 3);
-        column1.setInt(3, 4);
-        column1.setInt(4, 5);
-
-        val column2 = recordBatchBuilder.getColumnBuilder(1);
-        column2.setInt(0, 2);
-        column2.setInt(1, 3);
-        column2.setInt(2, 4);
-        column2.setInt(3, 5);
-        column2.setInt(4, 6);
-
-        recordBatchBuilders.add(recordBatchBuilder);
-        tableBuilder = new TableBuilder(client, schemaBuilder, recordBatchBuilders);
-        ObjectMeta meta = tableBuilder.seal(client);
-        client.persist(meta.getId());
-        client.putName(meta.getId(), "hive_example");
-        System.out.println("Table ID: " + meta.getId().value());
-        System.out.println("Prepare data done");
-    }
 }
-// how to get data from vineyard ?????
